Added promotions calculation functions.

+++ test/com/simonrobins/pe/TestPromotionEngine.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 class TestPromotionEngine {
 	@Test
 	void testScenarioA() {
-		Double cost = PromotionEngine.calculateOrderCost("ABC");
+		Integer cost = PromotionEngine.calculateOrderCost("ABC");
 		assertEquals(100, cost);
 		cost = PromotionEngine.calculateOrderCost("CBA");
 		assertEquals(100, cost);
@@ -18,7 +18,7 @@ class TestPromotionEngine {
 
 	@Test
 	void testScenarioB() {
-		Double cost = PromotionEngine.calculateOrderCost("AAAAABBBBBC");
+		Integer cost = PromotionEngine.calculateOrderCost("AAAAABBBBBC");
 		assertEquals(370, cost);
 		cost = PromotionEngine.calculateOrderCost("CBBBBBAAAAA");
 		assertEquals(370, cost);
@@ -26,7 +26,7 @@ class TestPromotionEngine {
 
 	@Test
 	void testScenarioC() {
-		Double cost = PromotionEngine.calculateOrderCost("AAABBBBBCD");
+		Integer cost = PromotionEngine.calculateOrderCost("AAABBBBBCD");
 		assertEquals(280, cost);
 		cost = PromotionEngine.calculateOrderCost("DCBBBBBAAA");
 		assertEquals(280, cost);
